requirements.txt
pyyaml==6.0.2
watchdog==5.0.3
